накидал каркас
