+++ models/__init__.py
@@ -1,0 +1,2 @@
+from .author_model import *
+from .post_model import *
