+++ models/author_model.py
@@ -1,0 +1,4 @@
+class AuthorPage:
+    @classmethod
+    def parse(cls, soup):
+        ...
