+++ models/post_model.py
@@ -1,0 +1,4 @@
+class PostPage:
+    @classmethod
+    def parse(cls, soup):
+        ...
