+++ my_task/__init__.py
@@ -1,0 +1,1 @@
+from .class_task import *
