+++ my_task/base_task.py
@@ -1,0 +1,7 @@
+class BaseTask:
+
+    def save(self):
+        ...
+
+    def get_hrefs(self):
+        ...
