+++ requirements.txt
@@ -1,2 +1,3 @@
-requests=2.27
-beautifulsoup4=4.10.0
+bs4==0.0.1
+beautifulsoup4==4.10.0
+requests==2.27.1
